refactor(fresco): extract terminal napi types into their own module
